CLI dataset reuse fix; profiles sized to measured step time; ruff clean for CI
- cli: a profile reuses only its own `<profile>-auto` dataset; the demo run
  had silently trained on the 24-document smoke set (job cancelled and
  marked failed with the reason).
- train profiles: demo 100 steps x accum 4 at 896 px (~35 min), overnight
  700 steps x accum 8 at 1024 px (~8 h), from ~6 s per micro-batch measured.
- ruff: 27 findings CI had been failing on (long literals, ambiguous glyphs,
  pairwise/unpacking idioms) fixed; local runs now gate on "All checks passed".

+++ packages/ml/src/ledgerlens_ml/synth.py
@@ -324,7 +324,7 @@ def generate_one(seed: int, *, degrade: float | None = None) -> SynthDoc:
     items_boxes: list[dict[str, list[float]]] = []
     subtotal = Decimal(0)
     for desc, unit, qty in chosen:
-        qty_v = max(1, int(round(qty * rnd.uniform(0.6, 1.6))))
+        qty_v = max(1, round(qty * rnd.uniform(0.6, 1.6)))
         unit_v = Decimal(str(unit)) * Decimal(str(rnd.choice([1, 1, 1.05, 0.95])))
         unit_v = unit_v.quantize(Decimal("0.01"), rounding=ROUND_HALF_UP)
         amount = (unit_v * qty_v).quantize(Decimal("0.01"), rounding=ROUND_HALF_UP)
